Add files via upload
Using tk inter module,saving information using .csv file,error and exception handling I created a small application where you can learn French. When it run the card will be shown with french word.After 3 seconds of delay the card face will be switched with corresponding english word. User will mark if the word is known or unknown.If it is a known word, it will be removed from csv file, if not it will be keept until user knows him.